requirements.txt
fastapi
uvicorn
python-multipart
openai>=1.12.0
python-dotenv
